CPA-278 split huge views.py of visualization module
* [x] Created views folder with __init__.py and imports
* [x] Created lib folder for common used functions
* [x] added new rule for ignore unused variable for __init__.py in flake8 config
* [x] Split country app views to multiple files

closed #278
