+++ country/views/__init__.py
@@ -1,0 +1,11 @@
+from .buyer_view import BuyerView
+from .country_view import CountryView
+from .data_edit_view import DataEditView
+from .data_import_view import DataImportView
+from .data_validate_view import DataValidateView
+from .delete_data_set_view import DeleteDataSetView
+from .language_view import LanguageView
+from .overall_stat_summary_view import OverallStatSummaryView
+from .supplier_view import SupplierView
+from .tender_pagination import TenderPagination
+from .tender_view import TenderView
