+++ setup.cfg
@@ -4,6 +4,9 @@
 exclude = migrations
 max-line-length = 119
 extend-ignore = E203, W503
+per-file-ignores =
+    # imported but unused
+    __init__.py: F401
 
 [isort]
 skip = migrations
